rag_study_app/backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.7.0
python-multipart>=0.0.7
requests>=2.31.0
pymongo
faiss-cpu==1.7.4
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.1
rich>=13.6.0
gunicorn>=22.1.0
google-generativeai

